+++ handlers/MapLibreMapBuilder.Windows.cs
@@ -42,5 +42,6 @@ public partial class MapLibreMapBuilder : IMapLibreMapOptionsSink
     public void SetAttributionButtonMargins(int x, int y)      { }
     public void SetShowNavigationControls(bool show)            { }
     public void SetShowAttributionControl(bool show, string? customAttribution) { }
+    public void SetShowGpsControl(bool show)                    { }
 }
 #endif