+++ handlers/MapLibreMapHandler.MaciOS.cs
@@ -24,8 +24,8 @@ public partial class MapLibreMapHandler : ViewHandler<MapLibreMap, UIView>
         _controller.OnCameraMoveReceived            += VirtualView.OnCameraMove;
         _controller.OnCameraIdleReceived            += VirtualView.OnCameraIdle;
         _controller.OnCameraTrackingChangedReceived += VirtualView.OnCameraTrackingChanged;
-        _controller.OnMapClickReceived              += VirtualView.OnMapClick;
-        _controller.OnMapLongClickReceived          += VirtualView.OnMapLongClick;
+        _controller.OnMapClickReceived              += (ll, sx, sy) => VirtualView.OnMapClick(ll, sx, sy);
+        _controller.OnMapLongClickReceived          += (ll, sx, sy) => VirtualView.OnMapLongClick(ll, sx, sy);
         _controller.OnUserLocationUpdateReceived    += VirtualView.OnUserLocationUpdate;
 
         return _controller.View;
@@ -48,4 +48,5 @@ public partial class MapLibreMapHandler : ViewHandler<MapLibreMap, UIView>
     public void UpdateAttributionButtonMargins(int?[]? margin) { if (margin?.Length >= 2 && margin[0] != null && margin[1] != null) _controller.SetAttributionButtonMargins(margin[0]!.Value, margin[1]!.Value); }
     public void UpdateShowNavigationControls(bool show) => _controller.SetShowNavigationControls(show);
     public void UpdateShowAttributionControl(bool show, string? customAttribution) => _controller.SetShowAttributionControl(show, customAttribution);
+    public void UpdateShowGpsControl(bool show)         => _controller.SetShowGpsControl(show);
 }